+++ CQRS.Application/Commands/NovaPessoaCommand/NovaPessoaCommandInput.cs
@@ -4,13 +4,13 @@ namespace CQRS.Application.Commands.NovaPessoaCommand
 {
     public class NovaPessoaCommandInput : CommandInput<NovaPessoaCommandResult>
     {
-        public NovaPessoaCommandInput(string nome, ushort idade)
+        public NovaPessoaCommandInput(string nome, byte idade)
         {
             Nome = nome;
             Idade = idade;
         }
 
         public string Nome { get; private set; }
-        public ushort Idade { get; private set; }
+        public byte Idade { get; private set; }
     }
 }
